Collection handling + query update

+++ database/migrations/2018_10_11_115045_add_building_id_to_user_progresses.php
@@ -6,39 +6,33 @@ use Illuminate\Support\Facades\Schema;
 
 class AddBuildingIdToUserProgresses extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::table('user_progresses', function (Blueprint $table) {
-            $table->integer('building_id')->unsigned()->nullable()->after('step_id');
-            $table->foreign('building_id')->references('id')->on('buildings')->onDelete('cascade');
-        });
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::table('user_progresses', function (Blueprint $table) {
+			if (!Schema::hasColumn('user_progresses', 'building_id')) {
+				$table->integer( 'building_id' )->unsigned()->nullable()->after( 'step_id' );
+				$table->foreign( 'building_id' )->references( 'id' )->on( 'buildings' )->onDelete( 'cascade' );
+			}
+		});
 
-        $progresses = DB::table('user_progresses')->get();
-        foreach($progresses as $progress){
-            $building = \DB::table('buildings')->where('user_id', '=', $progress->user_id)->get();
-            DB::table('user_progresses')->update([ 'building_id' => $building->id ])->where('user_id', '=', $progress->user_id);
-        }
+		$progresses = DB::table('user_progresses')->get();
+		foreach($progresses as $progress){
+			$building = \DB::table('buildings')->where('user_id', '=', $progress->user_id)->first();
+			DB::table('user_progresses')->where('user_id', '=', $progress->user_id)->update([ 'building_id' => $building->id ]);
+		}
+	}
 
-        /*
-        foreach (\App\Models\UserProgress::all() as $userProgress) {
-            $building = \App\Models\Building::find($userProgress->user_id);
-            $userProgress->building_id = $building->id;
-            $userProgress->save();
-        }
-        */
-    }
-
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+	}
 }
